requirements.txt
hyperliquid-python-sdk>=0.1.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
python-telegram-bot>=20.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
structlog>=24.1.0
numpy>=1.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
ruff>=0.4.0
mypy>=1.10.0
types-PyYAML>=6.0.12
